// frontend/src/api/models/groupDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GroupMemberDto } from './groupMemberDto';
import type { PlaylistSummaryDto } from './playlistSummaryDto';

export interface GroupDto {
  id?: string;
  name?: string;
  aboutUs?: string;
  isPublic?: boolean;
  createdAt?: string;
  updatedAt?: string;
  members?: GroupMemberDto[];
  playlists?: PlaylistSummaryDto[];
}