// frontend/src/api/models/groupMemberDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface GroupMemberDto {
  id?: string;
  userId?: string;
  username?: string;
  displayName?: string;
  avatarUrl?: string;
  isAdmin?: boolean;
  canEditInfo?: boolean;
  canManagePlaylists?: boolean;
  canInviteMembers?: boolean;
  canRemoveMembers?: boolean;
  status?: string;
  invitedAt?: string;
  acceptedAt?: string;
}